requirements.txt
textwrap
openai==1.45.1
pdfplumber==0.11.4
gradio==5.3.0
gradio_client==1.4.2
markdown-it-py==3.0.0
markdown2==2.5.1
weasyprint==63.0
nltk==3.8.1
textstat==0.7.4
spacy==3.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
textblob==0.18.0.post0
sentence-transformers==3.1.0
transformers==4.44.2
torch==2.4.1
torchvision==0.19.1